602 validation schemas in a file (#611)
* Put schemas in a separated files and update references

* Update chat.controller.js
