+++ backend/app/controllers/chat.controller.js
@@ -1,10 +1,11 @@
 const prisma = require('../../prisma/indexPrisma')
+const logger = require('../../logger')
+const { apiResponse, tokenUser } = require('../utils/utils')
 const {
-  apiResponse,
   conversationUsersSchema,
   conversationSchema,
   conversationUserSchema,
-} = require('../utils/utils')
+} = require('../utils/schemaValidation')
 
 async function createConversation(req, res) {
   try {
