+++ backend/app/middleware/validation.js
@@ -1,12 +1,10 @@
-module.exports = (schema) => {
-	return async (req, res, next) => {
-		try {
-			await schema.validateAsync(req.body);
-			next();
-		} catch (err) {
-			res.status(400).json({
-				msg: err.message,
-			});
-		}
-	};
-};
+module.exports = (schema) => async (req, res, next) => {
+  try {
+    await schema.validateAsync(req.body)
+    next()
+  } catch (err) {
+    res.status(400).json({
+      msg: err.message,
+    })
+  }
+}
